requirements.txt
google-genai
PyPDF2
python-dotenv
pydantic
fastapi
uvicorn
python-multipart